telegram_bot/requirements.txt
python-telegram-bot
python-dotenv
requests
watchdog
aiohttp